requirements.txt
hdbscan
librosa
matplotlib
numba
numpy
pandas
scikit_learn
scipy
seaborn
soundfile
umap_learn
cuda-python
